Finished Registration usecase

+++ library-server/src/main/java/de/h_da/library/registration/usecase/Registration.java
@@ -33,7 +33,7 @@ public interface Registration {
 	 *        ]
 	 * </pre>
 	 */
-	public int register(Customer customer);
+	public Long register(Customer customer);
 
 	/**
 	 * Modifies the data of an already registered customer
